+++ unstructured/ingest/main.py
@@ -6,6 +6,7 @@ import sys
 
 import click
 
+from unstructured.ingest.connector.github import GitHubConnector, SimpleGitHubConfig
 from unstructured.ingest.connector.reddit import RedditConnector, SimpleRedditConfig
 from unstructured.ingest.connector.s3_connector import S3Connector, SimpleS3Config
 from unstructured.ingest.doc_processor.generalized import initialize, process_document
@@ -80,6 +81,29 @@ class MainProcess:
     help="Connect to s3 without local AWS credentials.",
 )
 @click.option(
+    "--github-url",
+    default=None,
+    help='URL to GitHub repository, e.g. "https://github.com/Unstructured-IO/unstructured",'
+    ' or a repository owner/name pair, e.g. "Unstructured-IO/unstructured"',
+)
+@click.option(
+    "--github-access-token",
+    default=None,
+    help="A GitHub access token, see https://docs.github.com/en/authentication",
+)
+@click.option(
+    "--github-branch",
+    default=None,
+    help="The branch for which to fetch files from. If not given,"
+    " the default repository branch is used.",
+)
+@click.option(
+    "--github-file-glob",
+    default=None,
+    help="A comma-separated list of file globs to limit which types of files are accepted,"
+    " e.g. '*.html,*.txt'",
+)
+@click.option(
     "--subreddit-name",
     default=None,
     help='The name of a subreddit, without the "r\\", e.g. "machinelearning"',
@@ -148,6 +172,10 @@ class MainProcess:
 @click.option("-v", "--verbose", is_flag=True, default=False)
 def main(
     s3_url,
+    github_url,
+    github_access_token,
+    github_branch,
+    github_file_glob,
     subreddit_name,
     reddit_client_id,
     reddit_client_secret,
@@ -179,6 +207,21 @@ def main(
                 anonymous=s3_anonymous,
                 re_download=re_download,
                 preserve_downloads=preserve_downloads,
+                verbose=verbose,
+            ),
+        )
+    elif github_url:
+        doc_connector = GitHubConnector(  ***REMOVED*** type: ignore
+            config=SimpleGitHubConfig(
+                github_url=github_url,
+                github_access_token=github_access_token,
+                github_branch=github_branch,
+                github_file_glob=github_file_glob,
+                ***REMOVED*** defaults params:
+                download_dir=download_dir,
+                preserve_downloads=preserve_downloads,
+                output_dir=structured_output_dir,
+                re_download=re_download,
                 verbose=verbose,
             ),
         )